+++ tests/hal/test_nx_usb_properties.cpp
@@ -78,10 +78,10 @@ class USBPropertyTest : public ::testing::Test {
         size_t len = len_dist(rng);
 
         std::vector<uint8_t> data(len);
-        std::uniform_int_distribution<uint8_t> byte_dist(0, 255);
+        std::uniform_int_distribution<int> byte_dist(0, 255);
 
         for (size_t i = 0; i < len; i++) {
-            data[i] = byte_dist(rng);
+            data[i] = static_cast<uint8_t>(byte_dist(rng));
         }
 
         return data;
